Вынесла из компонента sort компонент dyas
