+++ src/components/sort.js
@@ -23,8 +23,6 @@ export const createSortTemplate = () => {
       </div>
 
       <span class="trip-sort__item  trip-sort__item--offers">Offers</span>
-    </form>
-
-    <ul class="trip-days"></ul>`
+    </form>`
   );
 };
